refactor(peagen): move GitVCS helpers to mirror_core
